+++ src/main/java/net/frozenblock/wilderwild/mixin/client/shrieker/LevelRendererMixin.java
@@ -22,6 +22,7 @@ import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.frozenblock.wilderwild.config.WWBlockConfig;
+import net.frozenblock.wilderwild.registry.WWParticleTypes;
 import net.frozenblock.wilderwild.registry.WWSounds;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.renderer.LevelRenderer;
@@ -54,15 +55,30 @@ public class LevelRendererMixin {
 	private Comparable<Boolean> wilderWild$shriekerGargle(Comparable<Boolean> original, int eventId, BlockPos pos, int data) {
 		if (this.level != null && WWBlockConfig.get().sculk.shriekerGargling) {
 			if (original.compareTo(true) == 0 || this.level.getFluidState(pos.above()).is(FluidTags.WATER)) {
+				double x = (double) pos.getX() + 0.5D;
+				double y = (double) pos.getY() + SculkShriekerBlock.TOP_Y;
+				double z = (double) pos.getZ() + 0.5D;
+
 				this.level.playLocalSound(
-					(double) pos.getX() + 0.5D,
-					(double) pos.getY() + SculkShriekerBlock.TOP_Y,
-					(double) pos.getZ() + 0.5D,
+					x,
+					y,
+					z,
 					WWSounds.BLOCK_SCULK_SHRIEKER_GARGLE,
 					SoundSource.BLOCKS,
 					2.0F,
 					0.6F + this.level.random.nextFloat() * 0.4F,
 					false
+				);
+
+				this.level.addParticle(
+					WWParticleTypes.SHRIEKER_BUBBLE_SPAWNER,
+					false,
+					x,
+					y,
+					z,
+					0D,
+					0D,
+					0D
 				);
 			}
 		}