+++ src/main/java/net/frozenblock/wilderwild/registry/WWParticleTypes.java
@@ -57,6 +57,7 @@ public final class WWParticleTypes {
 	public static final SimpleParticleType SCORCHING_FLAME = register("scorching_flame");
 	public static final SimpleParticleType LEAF_CLUSTER_SPAWNER = register("leaf_cluster");
 	public static final SimpleParticleType CHEST_BUBBLE_SPAWNER = register("chest_bubbles");
+	public static final SimpleParticleType SHRIEKER_BUBBLE_SPAWNER = register("shrieker_bubbles");
 	public static final SimpleParticleType BLUE_PEARLESCENT_HANGING_MESOGLEA = register("blue_pearlescent_hanging_mesoglea_drip");
 	public static final SimpleParticleType BLUE_PEARLESCENT_FALLING_MESOGLEA = register("blue_pearlescent_falling_mesoglea_drip");
 	public static final SimpleParticleType BLUE_PEARLESCENT_LANDING_MESOGLEA = register("blue_pearlescent_landing_mesoglea_drip");