+++ CHANGELOG.md
@@ -54,6 +54,7 @@ hi
 - Hanging Tendrils now play animations in sync with each other while being rendered as billboards.
 - Fixed an issue that prevented Hanging Tendrils from properly loading and saving Vibration data.
 - Fixed an issue that prevented Crabs from properly loading and saving Vibration data.
+- Particles created by gargling Sculk Shriekers are now handled by the client, using the new `wilderwild:shrieker_bubbles` particle.
 
 # Music
 - Added three new tracks by Willow/pictochats_.